requirements.txt
Flask==2.1.3
Werkzeug==2.0.3  # Compatible with Flask 2.1.3 as of my last knowledge update
Flask-SQLAlchemy==2.5.1  # Ensures compatibility with Flask 2.1.3
pymysql==1.0.2  # A commonly used MySQL driver for Python
python-dotenv==0.19.2  # For loading environment variables from .env files
bcrypt